[Add]データ型変更

+++ db/migrate/20220609105822_create_books.rb
@@ -1,10 +1,10 @@
 class CreateBooks < ActiveRecord::Migration[6.1]
   def change
     create_table :books, id: false do |t|
-      
+
       t.string :title
       t.string :author
-      t.integer :isbn, null: false, primary_key: true
+      t.bigint :isbn, null: false, primary_key: true
       t.string :sales_date
       t.string :publisher_name
       t.text :item_caption
